Detect and open all relevant ESLint config files
ESLint configuration files cascade, with precedence decided on closeness to the linted file in the hierarchy (OTOH, only the closest ignore file is used). We keep the default lookup mode to the closest files, as the presence alone is enough to check if ESLint has been configured at all, but add an option to walk the whole hierarchy.